-Updated Documentation -Added Tests -Changed dtype to float64 in pytorch for training with cpu -Adjusted covr to exclude functions exclusively used in shiny app

+++ inst/python/pytorch_autoencoder.py
@@ -254,7 +254,7 @@ log_dir=None, log_write_interval=10, log_top_value=0, log_top_total=1, log_top_m
   device=('cuda' if torch.cuda.is_available() else 'cpu')
   
   if device=="cpu":
-    dtype=float
+    dtype=torch.float64
     model.to(device,dtype=dtype)
   else:
     dtype=torch.double
@@ -381,7 +381,7 @@ def TeFeatureExtractorBatchExtract(model,dataset,batch_size):
   device=('cuda' if torch.cuda.is_available() else 'cpu')
   
   if device=="cpu":
-    dtype=float
+    dtype=torch.float64
     model.to(device,dtype=dtype)
   else:
     model.to(device,dtype=torch.double)
