+++ inst/python/pytorch_te_classifier.py
@@ -444,7 +444,8 @@ log_dir=None, log_write_interval=10, log_top_value=0, log_top_total=1, log_top_m
   device=('cuda' if torch.cuda.is_available() else 'cpu')
   
   if device=="cpu":
-    current_dtype=float
+    #current_dtype=float
+    current_dtype=torch.float64
     model.to(device,dtype=current_dtype)
   else:
     current_dtype=torch.double
